Fix/tzstats (#6)
Fix tzstats issues: As noted in the last PR, tzstats was in testing period and was therefore not set as default. We noticed that the staking balance given by tzstats should be retrieved from the snapshot call and not from the income call, otherwise (differently from tzscan) the staking balance of the payout cycle and not of the snapshot cycle will be taken as reference.
Furthermore, we added the possibility to get reward data of upcoming cycles based on the expected income.
HOWEVER, since tzstats does not allow commercial use only under previous agreement and that could be taken for granted, we will keep on not using tzstats as default. PRPC will be still the default, using tzstats should be explicitly set through the -P tzstats flag, and under personal guarantee in case of commercial use.