introduce "self-ideal" rewards like Backerei
I am the maintainer of backerei and in the process of depcrecating it.

Backerei payout operations are different than TRD: its default,
non-configurable behaviour is to pay rewards assuming perfect behaviour
of the baker doing the payouts, while still accounting for imperfect
behaviour of other bakers. This acts as a liveness guarantee and ensures
maximum payouts for delegators, but without overpaying for other baker's
downtime, like the current "ideal/expected" payout does.

This PR adds this capability to TRD.

Nomenclature
============

The current master branch has "ideal" and "actual" payouts. I initially
thought about naming this new payout type "ideal_self" or "self-ideal"
but it makes much more sense to do some renaming:

| old  | new    |
| ---- | ------ |
|ideal |expected|
|actual|actual  |
|      |ideal   |

The new names are more self-explanatory: "pay expected rewards" convey
that these rewards are not realized yet, while "pay ideal rewards"
includes the fees and extra rewards, which makes sense.

A drawback is that this will require a configuration change to current
users and will need to be release-noted. But, we expect most users of
the old-style "ideal/expected" rewards to be using them to pay out
future cycles. This is not possible with the new style "ideal" type,
because this reward type expect the cycle to actually have run, so these
users will have an error upon upgrade, which will force them to upgrade
their configuration from `rewards_type: ideal` to `rewards_type:
expected`.

I am open to any suggestions about better naming for this new model.

Implementation
==============

It only works with tzstats and tzkt API; at the moment, the TRD
implementation relying on Tezos Node RPC does not implement detailed
calculation of rewards, instead relying on unfreezing events to
determine the rewards. So, there is no way to add this feature without
implementing the detailed calculation itself. As a result, this type of
rewards will error out when attempted to run with a Tezos Node RPC
backend.

Both tzkt and tzstats "rewards" endpoint provide fields for missed
baking and endorsing rewards. When rewards_type.isIdeal is true, we add
these rewards to the "actual" rewards calculation, and this becomes our
total rewards.

While doing this, I observed that TRD was never paying out stolen block
rewards when used with the tzstats backend. This bug has been fixed.

I also observed that tzkt's rewards endpoints (but not tzstats) exposes
these fields:

* `missedOwnBlockFees`
* `missedExtraBlockFees`
* `uncoveredOwnBlockFees`
* `uncoveredExtraBlockFees`

These are the fees that the baker would have earned, had they baked the
blocks that they missed. We include them in the calculation because it
makes sense to do so. tzstats does not expose these lost fees, so there
is an ideal reward discrepancy between tzkt and tzstats backend when the
baker fails to bake one or more blocks in the cycle, with tzstats
backend reporting ideal payout lower than its true value. But this is
an edge case so it makes sense to enable ideal payouts with tzstats
backend anyway. We documented it.

Testing
=======

I tested different cycles for different bakers and verified that the
calculations were the same on tzkt/tzstats and that the actual rewards
matched what was showing on tzstats.