add ability to pay out frozen rewards with RPC
RPC has a helper to expose frozen rewards by cycle. This makes it
possible to pay out actual rewards before they unfreeze.

I am not sure if there is a specific reason this is not implemented? Is
it just that no one got to it, or is there a specific concern with these
values from RPC?