make default initial cycle take into account release_override
When no --initial_cycle and --release_override parameters are passed to
TRD,  assuming no previous payments have been
made, TRD will attempt to pay rewards for the cycle that most recently
unfroze.

If current cycle is 358, it will pay rewards for cycle 352.  This is normal
and expected behaviour.

The issue is that this behavior is not altered when --release_override
is used. If you want to pay rewards as soon as the cycle is
completed (--release_override -5) or as soon as the rights are
calculated (--release_override -11 and pay ideal rewards),
TRD STILL attemts to pay rewards for the cycle that most recently
unfroze when no --initial_cycle is given.

For example, if current cycle is 358:

* run TRD with `--release_override -5`, no prior payments (reports dir
  is empty) and no `--initial_cycle. It tries to pay rewards for cycle
  352, but the expected behavior is to attempt to pay cycle 357.
* run TRD with `--release_override -11`, no prior payments (reports dir
  is empty) and no `--initial_cycle. It tries to pay rewards for cycle
  352, but the expected behavior is to attempt to pay cycle 363.

This PR fixes this by taking into account release_override when
calculating the initial payment cycle when none is given.

Users explicitly passing a value to --initial_cycle are not impacted.
Users with previous succesful payments in their reports dir are not
affected. This only affects first-time users who set --release_override.

This is what I believe the user expects, but note that I am not
familiar with all the ways people run TRD, so reviewers are invited to
carefully consider the implications of this change based on their
interactions with users and their collected statistics.

In this PR I also edit some comments to make them clearer based on what
I discovered while reading code.