dry_run should not replay payments already done
Maybe again it's more a question than a bugfix, but I found myself doing
payments in -run_mode 2, and when I want to check before actually paying
with a `--dry_run` operation, it is replaying past payments. For
example, I was tyring to pay cycle 358, and I had already paid cycle 356
and 357, but the dry_run mode was retrying to pay cycles 356, 357, 358. I
gave --initial_cycle of 356.

Running in actual mode (no dry rnu) does not repay 356 and 357. It just
pays 358 as expected.

This PR makes dry run behave more like an actual payment run: it will
skip the payments that already have a "done.csv" file.