do not catch up on all late payments by default
The current behavior of TRD is to pay out every late payment at boot
time. So if my TRD did not run for a while, and I start it again, it
will pay everything it missed.

It may feel right, but it is not. Bakers almost certainly do not want to
run several payment cycles at once without interaction.

In regular TRD usage, you are not late. You want to pay the outstanding
payout for the cycle that has just been released.

If you happen to be late, you are probably not going to just restart the
software and call it a day.

Instead, you will be doing every payout manually, verifying after each
one and communicate with your delegates as you go, until the
situation is resolved.

So the current behavior on master is not useful. Most probably no one
wants this. But there is a chance that it happens by accident. Say, if
for some reason you lost most of your payout csv files until some date
in the past, and you do not pass the `--initial_cycle` parameter, you
will drain your payout account.

Therefore I propose to:

* not pay every payment since last success by default
* instead, assume `initial_payout_cycle=0` by default, which means that
  TRD will pay the recently released cycle only.
* you can still override this and pay out several unpaid cycles at once
  by passing a negative value to `initial_payout_cycle`, or an old
  cycle number.

I also removed some dead code that I found while doing this. There were
3 functions named `get_latest_report_file`, only one of which was used.