docker tag management
So far, we have only been pushing the "latest" tag.

We need to tag every docker image differently (release tags, commit tags)
to prevent unwanted upgrades of trd container.

We will still have a "latest" tag but we wouldn't recommend using it.

We are using the method described in github documentation:

https://docs.github.com/en/actions/guides/publishing-docker-images#publishing-images-to-docker-hub