fix issue with overwriting csv file containing pending overestimates
When the calculation csv file is first generated in early payout, the
overestimate is set as `pending`. Later, this field is overwritten with
the actual overestimate. I accidentally broke it during a code refactor,
and I didn't relaize until today. Today marks 5 cycles after I deployed
this PR for a production baker, so it's the first time a csv file that
says `pending` is actually overwritten. Before, csv files in the format
currently on master branch were being overwritten (without any
overestimate row at all).