fix "not in payment log" status in calculation
When adding #540 (advanced early payout) I made an assumption that
"payable" should be set to false if the adjustment is bigger than the
payout, resulting in an adjusted amount of 0.

It turns out, there are other reasons for adjusted_amount to be zero
(such as a payout redirect). The code handles it later. The positive
check affects this and results in unexpected description of the payout
in csv file.

Consequently I am removing this >0 check.