fix #678 - Fix single-shot mode always exiting in error
We added code in #653 to have TRD send a non-zero exit status when
anything wrong happens - including payout failures, or (relevant here)
keyboard interrupts.

However, it turns out that when using TRD in runmode -2, -3, -4, the
regular course of operations is to have the producer thread send a
KeyboardInterrupt when done... which is then treated as error!

This is done with `_thread.interrupt_main()`. It's bad: humans send
interrupts, programs should not. Instead, I'm suggesting that the
producer use `SIGUSR1` to trigger a normal interruption.

This triggers the same code path as today, just with an exit status of
SUCCESS. So I expect that the main thread will still wait for the
consumer, and it won't actually exit before payouts are done, if there
are any.

I'll deploy in a small baker and see if that's true. I have already
verified that when no payouts are due, exit code is now zero.

For the record, I still believe this entire producer/consumer logic
should be thrown away and we should simply do things in order
(calculate, then pay). I've ranted about this before (#491).