exit main thread with same exit code as payment_producer
https://github.com/tezos-reward-distributor-organization/tezos-reward-distributor/pull/679
introduced support for exit codes, so an alert can be sent in
single-shot mode when payouts fail for any reason.

However, it was crude, only supporting exit code 1.

The producer thread supports many exit codes.

In this case, there is a benign issue where tzkt returns "not synced"
and therefore payouts fail, but this is likely temporary and will pass
at next try, so there is no need to alert. But, currently it's not
possible to behave differently based on the exit code because it's
always 0 or 1.

An ugly solution is to save the exit code of the child thread in a file,
then read it in the main thread. That's what I am doing here. I remain
convinced that the entire thread architecture needs to go away, and we
need to make TRD single threaded again, but that's for another day.

Also:
* change the exit code of misconfigured provider to GENERAL_ERROR
  because it's not really a provider error,
* change the help to remove old providers that we don't support anymore